Rename `position` to `pos`
