+++ pumpkin-protocol/src/client/play/level_event.rs
@@ -1,6 +1,6 @@
 use pumpkin_data::packet::clientbound::PLAY_LEVEL_EVENT;
 use pumpkin_macros::client_packet;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use serde::Serialize;
 
 #[derive(Serialize)]
