+++ pumpkin-protocol/src/client/play/login.rs
@@ -1,5 +1,5 @@
 use pumpkin_data::packet::clientbound::PLAY_LOGIN;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 
 use pumpkin_macros::client_packet;
 use serde::Serialize;
