+++ pumpkin-protocol/src/client/play/open_sign_editor.rs
@@ -1,6 +1,6 @@
 use pumpkin_data::packet::clientbound::PLAY_OPEN_SIGN_EDITOR;
 use pumpkin_macros::client_packet;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use serde::Serialize;
 
 #[derive(Serialize)]
