+++ pumpkin-protocol/src/client/play/respawn.rs
@@ -1,6 +1,6 @@
 use pumpkin_data::packet::clientbound::PLAY_RESPAWN;
 use pumpkin_macros::client_packet;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use serde::Serialize;
 
 use crate::{VarInt, codec::identifier::Identifier};
