+++ pumpkin-protocol/src/server/play/pick_item.rs
@@ -1,6 +1,6 @@
 use pumpkin_data::packet::serverbound::{PLAY_PICK_ITEM_FROM_BLOCK, PLAY_PICK_ITEM_FROM_ENTITY};
 use pumpkin_macros::server_packet;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use serde::Deserialize;
 
 #[derive(Deserialize)]
