+++ pumpkin-protocol/src/server/play/update_sign.rs
@@ -1,7 +1,7 @@
 use bytes::Buf;
 use pumpkin_data::packet::serverbound::PLAY_SIGN_UPDATE;
 use pumpkin_macros::server_packet;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 
 use crate::{
     ServerPacket,
