+++ pumpkin/src/block/blocks/chest.rs
@@ -7,7 +7,7 @@ use pumpkin_data::{
 use pumpkin_inventory::{Chest, OpenContainer};
 use pumpkin_macros::pumpkin_block;
 use pumpkin_protocol::{client::play::CBlockAction, codec::var_int::VarInt};
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::registry::{Block, get_block};
 
 use crate::{
