+++ pumpkin/src/block/blocks/crafting_table.rs
@@ -7,7 +7,7 @@ use pumpkin_data::item::Item;
 use pumpkin_data::screen::WindowType;
 use pumpkin_inventory::{CraftingTable, OpenContainer};
 use pumpkin_macros::pumpkin_block;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::registry::Block;
 
 #[pumpkin_block("minecraft:crafting_table")]
