+++ pumpkin/src/block/blocks/furnace.rs
@@ -5,7 +5,7 @@ use pumpkin_data::item::Item;
 use pumpkin_data::screen::WindowType;
 use pumpkin_inventory::Furnace;
 use pumpkin_macros::pumpkin_block;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::registry::Block;
 
 use crate::{block::pumpkin_block::PumpkinBlock, server::Server};
