+++ pumpkin/src/block/blocks/mod.rs
@@ -1,6 +1,6 @@
 use pumpkin_data::screen::WindowType;
 use pumpkin_inventory::{Container, OpenContainer};
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::registry::Block;
 
 use crate::{entity::player::Player, server::Server};
