+++ pumpkin/src/block/mod.rs
@@ -18,7 +18,7 @@ use properties::{
 };
 use pumpkin_data::entity::EntityType;
 use pumpkin_data::item::Item;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_util::math::vec3::Vec3;
 use pumpkin_world::block::registry::{Block, State};
 use pumpkin_world::item::ItemStack;
