+++ pumpkin/src/block/properties/layers.rs
@@ -2,7 +2,7 @@ use crate::world::World;
 use async_trait::async_trait;
 use pumpkin_macros::block_property;
 use pumpkin_protocol::server::play::SUseItemOn;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::{
     BlockDirection,
     registry::{Block, State},
