+++ pumpkin/src/block/properties/signal_fire.rs
@@ -2,7 +2,7 @@ use crate::world::World;
 use async_trait::async_trait;
 use pumpkin_macros::block_property;
 use pumpkin_protocol::server::play::SUseItemOn;
-use pumpkin_util::math::{position::BlockPos, vec3::Vec3};
+use pumpkin_util::math::{pos::BlockPos, vec3::Vec3};
 use pumpkin_world::block::{BlockDirection, registry::Block};
 
 use super::{BlockProperties, BlockProperty, BlockPropertyMetadata, Direction};
