+++ pumpkin/src/block/pumpkin_block.rs
@@ -6,7 +6,7 @@ use async_trait::async_trait;
 use pumpkin_data::item::Item;
 use pumpkin_inventory::OpenContainer;
 use pumpkin_protocol::server::play::SUseItemOn;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::BlockDirection;
 use pumpkin_world::block::registry::Block;
