+++ pumpkin/src/block/registry.rs
@@ -5,7 +5,7 @@ use crate::world::World;
 use pumpkin_data::item::Item;
 use pumpkin_inventory::OpenContainer;
 use pumpkin_protocol::server::play::SUseItemOn;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::BlockDirection;
 use pumpkin_world::block::registry::Block;
 use std::collections::HashMap;
