+++ pumpkin/src/entity/mod.rs
@@ -21,7 +21,7 @@ use pumpkin_protocol::{
 use pumpkin_util::math::{
     boundingbox::{BoundingBox, EntityDimensions},
     get_section_cord,
-    position::BlockPos,
+    pos::BlockPos,
     vec2::Vec2,
     vec3::Vec3,
     wrap_degrees,
