+++ pumpkin/src/item/pumpkin_item.rs
@@ -2,7 +2,7 @@ use crate::entity::player::Player;
 use crate::server::Server;
 use async_trait::async_trait;
 use pumpkin_data::item::Item;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::registry::Block;
 
 pub trait ItemMetadata {
