+++ pumpkin/src/item/registry.rs
@@ -1,7 +1,7 @@
 use crate::entity::player::Player;
 use crate::server::Server;
 use pumpkin_data::item::Item;
-use pumpkin_util::math::position::BlockPos;
+use pumpkin_util::math::pos::BlockPos;
 use pumpkin_world::block::registry::Block;
 use std::collections::HashMap;
 use std::sync::Arc;
